moves `Appointments` to a component, removes the page
