+++ frontend/components/Appointments/index.ts
@@ -1,0 +1,1 @@
+export * from './Appointments';
